[language] Move simple transactions to a separate crate to get rid of the dependencies.
Previously all the simple transactions resides in the same crate as the genesis transaction, which was bringing in tons of unnecessary dependencies, as to generate a genesis we would need the entire VM. This diff separates out those code so that the client will no longer have to import the dependency for VM while encoding the transactions. This should also remove the node software's dependency on the move ir_to_bytecode_compiler.
