+++ benchmark/Cargo.toml
@@ -31,7 +31,7 @@ proto_conv = { path = "../common/proto_conv" }
 rusty-fork = "0.2.1"
 tools = { path = "../common/tools" }
 types = { path = "../types" }
-vm_genesis = { path = "../language/vm/vm_genesis" }
+transaction_builder = { path = "../language/transaction_builder" }
 
 [dev-dependencies]
 crypto = { path = "../crypto/crypto", features = ["testing"] }
