+++ benchmark/src/load_generator.rs
@@ -117,7 +117,7 @@ fn gen_mint_txn_request(
     faucet_account: &mut AccountData,
     receiver: &AccountAddress,
 ) -> Result<Request> {
-    let program = vm_genesis::encode_mint_script(receiver, FREE_LUNCH);
+    let program = transaction_builder::encode_mint_script(receiver, FREE_LUNCH);
     let signer = faucet_account
         .key_pair
         .as_ref()
@@ -133,7 +133,7 @@ fn gen_transfer_txn_request(
     wallet: &WalletLibrary,
     num_coins: u64,
 ) -> Result<Request> {
-    let program = vm_genesis::encode_transfer_script(&receiver, num_coins);
+    let program = transaction_builder::encode_transfer_script(&receiver, num_coins);
     gen_submit_transaction_request(program, sender, wallet)
 }
