+++ language/e2e_tests/src/tests/verify_txn.rs
@@ -14,6 +14,7 @@ use compiler::Compiler;
 use config::config::{NodeConfigHelpers, VMPublishingOption};
 use crypto::{ed25519::*, HashValue};
 use std::collections::HashSet;
+use transaction_builder::encode_transfer_script;
 use types::{
     test_helpers::transaction_test_helpers,
     transaction::{
@@ -23,7 +24,6 @@ use types::{
     vm_error::{StatusCode, StatusType, VMStatus},
 };
 use vm::gas_schedule::{self, GasAlgebra};
-use vm_genesis::encode_transfer_script;
 
 #[test]
 fn verify_signature() {
