+++ language/transaction_builder/Cargo.toml
@@ -6,20 +6,14 @@ license = "Apache-2.0"
 edition = "2018"
 
 [dependencies]
+config = { path = "../../config" }
+crypto = { path = "../../crypto/crypto" }
 failure = { path = "../../common/failure_ext", package = "failure_ext" }
+ir_to_bytecode = { path = "../compiler/ir_to_bytecode" }
+lazy_static = "1.3.0"
+stdlib = { path = "../stdlib" }
 types = { path = "../../types" }
 vm = { path = "../vm" }
-structopt = { version = "0.2.15", optional = true }
-serde_json = { version = "1.0.40", optional = true }
-proto_conv = { path = "../../common/proto_conv", optional = true }
 
 [dev-dependencies]
 types = { path = "../../types", features = ["testing"]}
-
-[features]
-build-binary = ["structopt", "serde_json", "proto_conv"]
-
-[[bin]]
-name = "transaction_builder"
-path = "src/bin/main.rs"
-required-features = ["build-binary"]
