+++ language/vm/vm_genesis/Cargo.toml
@@ -9,6 +9,7 @@ publish = false
 [dependencies]
 config = { path = "../../../config" }
 failure = { path = "../../../common/failure_ext", package = "failure_ext" }
+transaction_builder = { path = "../../transaction_builder"}
 ir_to_bytecode = { path = "../../compiler/ir_to_bytecode" }
 crypto = { path = "../../../crypto/crypto" }
 stdlib = { path = "../../stdlib" }
